take into consideration ext-mon

+++ scripts/dns/delete_dns_entries.sh
@@ -120,7 +120,7 @@ while read -r line; do
         stringarray=($line)
         TXT_RECORD_EXISTS=$(az network dns record-set txt show -g ${RESOURCE_GROUP} -z ${DNS_ZONE} -n ${stringarray[1]} 2>&1)
 
-        if [[ ""${TXT_RECORD_EXISTS}"" == *"Not Found"* ]]; then
+        if [[ ""${TXT_RECORD_EXISTS}"" == *"Not Found"* ]] && [[ ""${stringarray[1]}"" != "*.ext-mon" ]]; then
             $(az network dns record-set a delete -y -g ${RESOURCE_GROUP} -z ${DNS_ZONE} -n ${stringarray[1]})
             echo "Deleted ${stringarray[1]}"
         fi
